chore: write functions for comment functions
