+++ src/routes/[name]/games/[id]/+page.server.ts
@@ -1,7 +1,11 @@
+import { add_comments } from '$lib/server/games/comments/add_comments';
+import { add_reply } from '$lib/server/games/comments/add_reply';
+import { delete_comment } from '$lib/server/games/comments/delete_comment';
+import { delete_reply } from '$lib/server/games/comments/delete_reply';
+import { Game_Model } from '$lib/server/games/models';
 import { User_Model } from '$lib/server/user/models';
-
 import type { Types } from 'mongoose';
-import type { PageServerLoad } from '../$types';
+import type { Actions, PageServerLoad } from './$types';
 
 function removeIdFromGame(
 	game: Types.Subdocument<Types.ObjectId> & {
@@ -42,4 +46,42 @@ export const load: PageServerLoad = async (event: any) => {
 	const game = removeIdFromGame(gameWithId);
 
 	return { param_name, param_id, game };
+};
+
+export const actions: Actions = {
+	add_comment: async (event) => {
+		const email = event.cookies.get('email');
+		const formData = await event.request.formData();
+		const comment = formData.get('comment') as string;
+		const user = await User_Model.findOne({ 'user.email': email });
+		let game = await Game_Model.findOne({ id: event.params.id });
+		let user_game = user?.games.find((game) => game.id === event.params.id);
+
+		await add_comments(user, game, comment, user_game);
+	},
+	edit_comment: async (event) => {},
+	delete_comment: async (event) => {
+		const email = event.cookies.get('email') as string;
+		const formData = await event.request.formData();
+		const comment_id = formData.get('id') as string;
+
+		await delete_comment(email, event, comment_id);
+	},
+	add_reply: async (event) => {
+		const email = event.cookies.get('email') as string;
+		const formData = await event.request.formData();
+		const comment_id = formData.get('comment_id') as string;
+		const reply = formData.get('reply') as string;
+
+		await add_reply(event, email, comment_id, reply);
+	},
+	edit_reply: async (event) => {},
+	delete_reply: async (event) => {
+		const email = event.cookies.get('email') as string;
+		const formData = await event.request.formData();
+		const comment_id = formData.get('comment_id') as string;
+		const reply_id = formData.get('reply_id') as string;
+
+		await delete_reply(event, email, comment_id, reply_id);
+	}
 };
