+++ src/routes/[name]/games/[id]/+page.svelte
@@ -1,8 +1,16 @@
 <script lang="ts">
 	import FourWinningTable from '$lib/components/Archive/FourWinning_table.svelte';
+	import Comments from '$lib/components/Comments.svelte';
+	import { faArrowLeft } from '@fortawesome/free-solid-svg-icons';
+	import Fa from 'svelte-fa';
 	import { _, isLoading } from 'svelte-i18n';
 
 	export let data;
+	export let form: FormData;
+
+	function go_back() {
+		location.href = `/${data.param_name}/games`;
+	}
 </script>
 
 <svelte:head>
@@ -20,6 +28,16 @@
 {#if $isLoading}
 	<p>Loading...</p>
 {:else}
+	<button class="back" on:click={go_back}><Fa icon={faArrowLeft}></Fa></button>
+
+	{#if form?.error}
+		<p class="error">{form.status} - {form.error}</p>
+	{/if}
+
+	{#if form?.message}
+		<p class="success">{form.status} - {form.message}</p>
+	{/if}
+
 	<h1>
 		{#if data.game}
 			{$_('public_game', {
@@ -34,10 +52,32 @@
 	</h1>
 
 	{#if data.game}
-		<p class="data">{data.game.data}</p>
-
-		{#if data.game.teams.includes('4winning_')}
-			<FourWinningTable measurement_units="yards" data={data.game.data} />
+		{#if data?.game.owner}
+			<h2>{$_('owner')}</h2>
+			<p>{data.game.owner}</p>
 		{/if}
+
+		{#if data?.game.type}
+			<h2>{$_('type')}</h2>
+			<p>{data.game.type}</p>
+		{/if}
+
+		{#if data.game}
+			<p class="data">{data.game.data}</p>
+
+			{#if data?.game?.teams.includes('4winning_')}
+				<FourWinningTable measurement_units="yards" data={data.game.data} />
+			{/if}
+		{/if}
+
+		<Comments comments={data.game?.comments}></Comments>
 	{/if}
 {/if}
+
+<style>
+	.back {
+		background-color: transparent;
+		border: none !important;
+		outline: none !important;
+	}
+</style>
